requirements.txt
cmsplugin-filer==1.0.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.5
django-appconf==1.0.1
django-classy-tags==0.7.0
django-cms==3.2.0
django-filer==1.1.0
django-filter==0.13.0
django-formtools==1.0
django-mptt==0.8.0
django-polymorphic==0.8.1
django-reversion==1.8.7
django-sekizai==0.9.0
Django-Select2==4.3.2
django-toolbelt==0.0.1
django-treebeard==3.0
djangocms-admin-style==1.0.8
djangocms-column==1.6.0
djangocms-googlemap==0.4.0
djangocms-inherit==0.1.1
djangocms-installer==0.8.5
djangocms-link==1.7.1
djangocms-style==1.6.0
djangocms-text-ckeditor==2.8.1
djangorestframework==3.3.3
easy-thumbnails==2.3
gunicorn==19.3.0
html5lib==1.0b8
Markdown==2.6.6
mechanize==0.2.5
Pillow==3.1.0
postgres==2.2.1
psycopg2==2.6.1
pytz==2015.7
requests==2.9.1
six==1.10.0
static3==0.6.1
tzlocal==1.2
Unidecode==0.4.19
virtualenv==13.1.2
whitenoise==2.0.4
